+++ AmperfyTests/Cases/SongTest.swift
@@ -113,7 +113,7 @@ class SongTest: XCTestCase {
     }
     
     func testSyncWave() {
-        let testWaveId: Int16 = 987
+        let testWaveId: Int = 987
         let testWave = storage.createSyncWave()
         testWave.id = testWaveId
         testSong.syncInfo = testWave
